+++ src/adapters/BlockchainAdapterMem.ts
@@ -36,7 +36,7 @@ export class BlockchainAdapterMem implements IBlockchainAdapter {
       versionId: string;
       resourceType?: string;
       resourceId?: string;
-      fullUrl?: string;
+      tags?: Array<{ id: string; system?: string; code?: string; version?: string; userSelected?: boolean }>;
     }>,
     channel: string,
     chaincode: string,