+++ src/adapters/BlockchainAdapterMulti.ts
@@ -31,7 +31,7 @@ export class BlockchainAdapterMulti implements IBlockchainAdapter {
       versionId: string;
       resourceType?: string;
       resourceId?: string;
-      fullUrl?: string;
+      tags?: Array<{ id: string; system?: string; code?: string; version?: string; userSelected?: boolean }>;
     }>,
     channel: string,
     chaincode: string,